project complete

+++ src/Pages/Apply/Apply.jsx
@@ -1,21 +1,34 @@
 import Header from "../../components/Header/Header";
 import { useEffect, useState } from "react";
-import { useLocation } from 'react-router-dom'
+import { useLocation } from "react-router-dom";
+import SingleApplied from "../../components/SingleApplied/SingleApplied";
 
-export default function Applied() 
-    {
-        
-        let items = JSON.parse(localStorage.getItem("items"))
+export default function Applied() {
+
+    let items = JSON.parse(localStorage.getItem("items"))
         var existingEntries = JSON.parse(localStorage.getItem("allEntries"));
         if(existingEntries == null) existingEntries = [];
-        console.log(existingEntries);
-        existingEntries.push(items);
-        localStorage.setItem("allEntries", JSON.stringify(existingEntries));
-      
+        
+        if(!existingEntries.some(item => item.id === items.id)){
+            existingEntries.push(items);
+            localStorage.setItem("allEntries", JSON.stringify(existingEntries));
+        }
+  return (
 
-        return (
-            <>
-                <Header/>
-            </>
-        );
-    }
+    
+
+    <>
+      <Header />
+      <div className="container">
+        <div className="row row-cols-2">
+        {
+            existingEntries.map((appliedJob) => (
+                <SingleApplied key={appliedJob.id} appliedJob={appliedJob} />
+              ))
+        }
+
+        </div>
+      </div>
+    </>
+  );
+}
